Create README.md
# Hands-on Python Workshop

This repostory contains material for the STROBE seminar that will be held on Feb. 26th 2018
at UCLA, and in videoconference with UC Beckley, CU Boulder and UC Irvine.

During this workshop we will get the hands dirty analyzing the blinking trajectories of a 
single-molecule fluorescence microscopy video.
We use the [Jupyter Notebook](http://jupyter.org/) environment with Python 3.6
and common scientific packages (Numpy, Scipy, Pandas, Matplotlib and Seaborn).

## Installation

To save time, partecipants will use a dedicated JupyterHub server with all the software 
already installed. The only program needed is a web browser to connect to the server.
However, we will shut-down the server a few days after the workshop. So, if you want to 
run again the analysis you need to install the software on your own computer
and download the data and notebooks from this repository.

To install the Python and Jupyter on your own computer, we recommend installing the
[Anaconda Scientific Python Distribution](https://www.anaconda.com/) 
which includes the important packages for scientific computing with Python. 
All the software is free and open source and multi-platform (Windows, Linux, Mac).

## Content

This repository contains notebooks and data files used for the workshop.

Please get familiar with Numpy indexing before the workshop by reading and running
the notebook:

- [Numpy Primer](https://github.com/tritemio/Jupyter-Python-Workshop-STROBE-2018-02-26/blob/master/Numpy%20primer.ipynb)

You will find the same notebook in the Jupyter sever after you login.